✨ Home page: Book list to be fixed and componentised
